Revert "Ultra-Realistic 3D EV Model Enhancement with Interactive Features and Inventory Integration"
